WIP: main app compiles, fixing test errors

+++ app/uk/gov/hmrc/plasticpackagingtaxreturns/config/AppConfig.scala
@@ -53,7 +53,7 @@ class AppConfig @Inject() (config: Configuration, servicesConfig: ServicesConfig
 
   val auditingEnabled: Boolean   = config.get[Boolean]("auditing.enabled")
   val graphiteHost: String       = config.get[String]("microservice.metrics.graphite.host")
-  val dbTimeToLiveInSeconds: Int = config.get[Int]("mongodb.timeToLiveInSeconds")
+  val dbTimeToLiveInSeconds: Long = config.get[Long]("mongodb.timeToLiveInSeconds")
 
   val eisEnvironment: String = config.get[String]("eis.environment")
